@@ -7,68 +7,71 @@
     "404": "cp out/404/index.html out/404.html",
     "test": "jest --passWithNoTests",
     "dev": "next dev",
-    "prebuild": "yarn generate",
-    "build": "next build && next export",
-    "release": "yarn build",
+    "prebuild": "pnpm install",
+    "build": "next build",
     "start": "next start",
     "analyze": "ANALYZE=true next build",
-    "prelint": "yarn install",
-    "lint": "next lint",
+    "lint": "eslint --ext .ts .tsx src",
+    "lint-fix": "eslint --ext .ts .tsx src --fix",
+    "style-check": "prettier src",
+    "format": "prettier --write src",
     "clean": "rimraf build",
-    "pretest": "yarn build",
-    "all": "yarn install && yarn lint && yarn test",
-    "postbuild": "yarn 404",
-    "serve": "http-server out -p 8080",
-    "generate": "typewind generate",
-    "pregenerate": "yarn install"
+    "pretest": "pnpm build",
+    "all": "pnpm install && pnpm lint && pnpm test",
+    "postbuild": "pnpm 404",
+    "prestyle-check": "pnpm install",
+    "preformat": "pnpm install",
+    "prelint": "pnpm style-check",
+    "prelint-fix": "pnpm format",
+    "serve": "http-server out -p 8080"
   },
   "dependencies": {
-    "@emotion/react": "*",
-    "@emotion/server": "*",
-    "@emotion/styled": "*",
-    "@mui/material": "*",
-    "i18next": "*",
-    "next": "^13",
-    "next-i18next": "*",
-    "next-language-detector": "*",
-    "react": "*",
-    "react-dom": "*",
-    "react-i18next": "*",
-    "web-vitals": "*"
+    "@emotion/cache": "^11.11.0",
+    "@emotion/react": "^11.11.1",
+    "@emotion/server": "^11.11.0",
+    "@emotion/styled": "^11.11.0",
+    "@mui/material": "^5.13.4",
+    "i18next": "^22.5.1",
+    "next": "^13.4.4",
+    "next-i18next": "^13.3.0",
+    "next-language-detector": "^1.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-i18next": "^12.3.1",
+    "web-vitals": "^3.3.2"
   },
   "devDependencies": {
-    "@mui/icons-material": "*",
-    "@next/bundle-analyzer": "^13",
-    "@swc/jest": "*",
-    "@types/jest": "*",
-    "@types/node": "*",
-    "@types/react": "*",
-    "@typescript-eslint/eslint-plugin": "*",
-    "@typescript-eslint/parser": "*",
-    "autoprefixer": "*",
-    "eslint": "*",
-    "eslint-config-airbnb-base": "*",
-    "eslint-config-airbnb-typescript": "*",
-    "eslint-config-import": "*",
-    "eslint-config-next": "^13",
-    "eslint-config-prettier": "*",
-    "eslint-import-resolver-typescript": "*",
-    "eslint-plugin-flowtype": "*",
-    "eslint-plugin-import": "*",
-    "eslint-plugin-jest": "*",
-    "eslint-plugin-jsx-a11y": "*",
-    "eslint-plugin-prettier": "*",
-    "eslint-plugin-react": "*",
-    "eslint-plugin-react-hooks": "*",
-    "http-server": "*",
-    "jest": "*",
-    "next-compose-plugins": "*",
-    "postcss": "*",
-    "prettier": "*",
-    "rimraf": "*",
-    "tailwindcss": "*",
-    "typescript": "*",
-    "typewind": "*"
+    "@mui/icons-material": "^5.11.16",
+    "@next/bundle-analyzer": "^13.4.4",
+    "@swc/jest": "^0.2.26",
+    "@types/jest": "^29.5.2",
+    "@types/node": "^20.2.5",
+    "@types/react": "^18.2.8",
+    "@typescript-eslint/eslint-plugin": "^5.59.9",
+    "@typescript-eslint/parser": "^5.59.9",
+    "autoprefixer": "^10.4.14",
+    "eslint": "^8.42.0",
+    "eslint-config-airbnb-base": "^15.0.0",
+    "eslint-config-airbnb-typescript": "^17.0.0",
+    "eslint-config-import": "^0.13.0",
+    "eslint-config-next": "^13.4.4",
+    "eslint-config-prettier": "^8.8.0",
+    "eslint-import-resolver-typescript": "^3.5.5",
+    "eslint-plugin-flowtype": "^8.0.3",
+    "eslint-plugin-import": "^2.27.5",
+    "eslint-plugin-jest": "^27.2.1",
+    "eslint-plugin-jsx-a11y": "^6.7.1",
+    "eslint-plugin-prettier": "^4.2.1",
+    "eslint-plugin-react": "^7.32.2",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "http-server": "^14.1.1",
+    "jest": "^29.5.0",
+    "next-compose-plugins": "^2.2.1",
+    "postcss": "^8.4.24",
+    "prettier": "^2.8.8",
+    "rimraf": "^5.0.1",
+    "tailwindcss": "^3.3.2",
+    "typescript": "^5.1.3"
   },
   "jest": {
     "transform": {
