@@ -1,13 +1,36 @@
-import { Box, Typography } from '@mui/material';
+import { Box, Link, Typography } from '@mui/material';
 import { Trans, useTranslation } from 'react-i18next';
 
 import Layout from '../../components/Layout';
-import StyledBox from 'components/styled/StyledBox';
-import CardStyleLink from 'components/styled/CardStyleLink';
 import { getStaticPaths, makeStaticProps } from 'lib/getStatic';
+import { ReactNode } from 'react';
 
+function CardStyleLink({ children, href, className }: { children: ReactNode, href: string, className?: string }) {
+  return (<Link href={href} sx={{
+    padding: '18px 18px 24px',
+    width: '220px',
+    textAlign: 'left',
+    color: '#434343',
+    border: '1px solid #9b9b9b',
+    textDecoration: 'none',
+    '&:hover, &.Mui-focusVisible, &.Mui-active': {
+      borderColor: '#067df7',
+    },
+    '& h3': {
+      margin: '0',
+      color: '#067df7',
+      fontSize: '18px',
+    },
+    '& p': {
+      margin: '0',
+      padding: '12px 0 0',
+      color: '#333',
+      fontSize: '13px',
+    },
+  }} className={className}>{children}</Link>);
+}
 
-const Home = (): JSX.Element => {
+function Home() {
   const { t } = useTranslation();
 
   return (
@@ -39,7 +62,14 @@ const Home = (): JSX.Element => {
           />
         </Typography>
 
-        <StyledBox>
+        <Box sx={{
+          maxWidth: '880px',
+          display: 'flex',
+          flexDirection: 'row',
+          justifyContent: 'space-around',
+          margin: '80px auto 40px',
+          textDecoration: 'none',
+        }}>
           <CardStyleLink href="https://nextjs.org/docs">
             <Typography component='h3'>{t('home:doc')} &rarr;</Typography>
             <Typography component='p'>{t('home:docDesc')}</Typography>
@@ -52,7 +82,7 @@ const Home = (): JSX.Element => {
             <Typography component='h3'>{t('home:examples')} &rarr;</Typography>
             <Typography component='p'>{t('home:examplesDesc')}</Typography>
           </CardStyleLink>
-        </StyledBox>
+        </Box>
       </Box>
     </Layout>
   );
